server/requirements.txt
Flask==3.1.1
mysql-connector==2.2.9
jproperties==2.1.2